Avoid another int64 casting

+++ pandas/core/window/rolling.py
@@ -2231,9 +2231,7 @@ class RollingGroupby(WindowGroupByMixin, Rolling):
         """
         # Ensure the object we're rolling over is monotonically sorted relative
         # to the groups
-        groupby_order = np.concatenate(
-            list(self._groupby.grouper.indices.values())
-        ).astype(np.int64)
+        groupby_order = np.concatenate(list(self._groupby.grouper.indices.values()))
         obj = obj.take(groupby_order)
         return super()._create_blocks(obj)
